Update FakeUniswap.sol
This contract has already been deployed on Rinkeby and tested using 'testSwap.sol'.  Format for using this can be found in 'testSwap.sol'.  Transfer Weth, and receive an amount of DAI/LINK depending on a fixed exchange rate.  This will be used to test out uniswap functionality, and multi-ERC20 portfolio contracts.